samples/python/agents/WeatherAgentADK/pyproject.toml
[project]
name = "a2a-sample-agent-adk"
version = "0.1.0"
description = "Sample Google weather agent." 
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "google-adk>=0.0.3",
    "google-genai>=1.9.0",
    "python-dotenv>=1.1.0",
     "litellm>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
common = { path = "../common", editable = true }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"